Added new functions
Added undo stack function
Added exit function
Added Indented print function
added write data to file function

+++ src/Crypto.h
@@ -20,20 +20,20 @@ class Cryptocurrency
 private:
 	string coinName;
 	string alg;
-	int supply;
+	long long supply;
 	int dateFound;
 	double price;
 	string founder;
 
 public:
 	Cryptocurrency();
-	Cryptocurrency(string n, string a, int s, int d, double p, string f);
+	Cryptocurrency(string n, string a, long long s, int d, double p, string f);
 	~Cryptocurrency();
 
 	// Setters
 	void setName(string n) { coinName = n; }
 	void setAlg(string a) { alg = a; }
-	void setSupply(int s) { supply = s; }
+	void setSupply(long long s) { supply = s; }
 	void setDate(int d) { dateFound = d; }
 	void setPrice(double p) { price = p; }
 	void setFounder(string f) { founder = f; }
@@ -41,7 +41,7 @@ public:
 	// Getters
 	string getName() const { return coinName; }
 	string getAlg() const { return alg; }
-	int getSupply() const { return supply; }
+	long long getSupply() const { return supply; }
 	int getDate() const { return dateFound; }
 	double getPrice() const { return price; }
 	string getFounder() const { return founder; }
